HW12: Renamed hw12_13_14_15_calendar to hw12_calendar
