+++ hw12_13_14_15_calendar/go.mod
@@ -1,4 +1,4 @@
-module github.com/FRiniZ/otus-go-hw-test/hw12_13_14_15_calendar
+module github.com/FRiniZ/otus-go-hw-test/hw12_calendar
 
 go 1.16
