+++ hw12_13_14_15_calendar/internal/app/app.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/FRiniZ/otus-go-hw-test/hw12_13_14_15_calendar/internal/storage"
+	"github.com/FRiniZ/otus-go-hw-test/hw12_calendar/internal/storage"
 )
 
 var ErrDateBusy = errors.New("date is busy")
