+++ hw12_13_14_15_calendar/internal/storage/memory/storage.go
@@ -3,7 +3,7 @@ package memorystorage
 import (
 	"sync"
 
-	"github.com/FRiniZ/otus-go-hw-test/hw12_13_14_15_calendar/internal/storage"
+	"github.com/FRiniZ/otus-go-hw-test/hw12_calendar/internal/storage"
 )
 
 type mapEvent map[int64]*storage.Event
