+++ hw12_13_14_15_calendar/internal/storage/memory/storage_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	storage "github.com/FRiniZ/otus-go-hw-test/hw12_13_14_15_calendar/internal/storage"
+	storage "github.com/FRiniZ/otus-go-hw-test/hw12_calendar/internal/storage"
 	faker "github.com/go-faker/faker/v4"
 	"github.com/stretchr/testify/require"
 )
